fixed unagro and pathfinding bugs

+++ Lab3/app/src/main/java/com/example/lab3/entities/GameObject.java
@@ -40,7 +40,7 @@ public abstract class GameObject {
         );
     }
 
-    public static double getMapDistBetweenObjects(GameObject obj1, GameObject obj2){
+    public static int getMapDistBetweenObjects(GameObject obj1, GameObject obj2){
         return Math.max(Math.abs(obj1.mapPosX - obj2.mapPosX), Math.abs(obj1.mapPosY - obj2.mapPosY));
     }
 }
